requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
pymongo==3.10.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
